buffer-cursor: create individual file for BufferCursor
